backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "backend"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Andres D. Molins", email = "amolinsdiaz@gmail.com" },
  { name = "Gerard Molina", email = "gerardmo98@gmail.com" },
  { name = "Reza Rahemtola", email = "contact@rezar.fr" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "fastapi==0.115.8",
  "pydantic==1.10.21",
  "uuid==1.30",
  "paramiko==3.5.0",
  "eciespy==0.4.3",
  "python-dotenv==1.0.1",
  "aleph-sdk-python==1.4.0",
]
optional-dependencies.all = [
  "fastapi[standard]",
]

[project.urls]
Documentation = "https://github.com/unknown/backend#readme"
Issues = "https://github.com/unknown/backend/issues"
Source = "https://github.com/unknown/backend"

[tool.hatch.version]
path = "src/backend/__about__.py"

[tool.hatch.build.targets.wheel]
packages = [ "src/backend" ]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.envs.default]
platforms = ["linux", "macos"]

[tool.hatch.envs.default.scripts]
backend = "fastapi run src/backend/main.py {args:--help}"

[tool.hatch.envs.testing]
type = "virtual"
system-packages = true
dependencies = [
  "coverage[toml]>=6.5",
  "pytest",
]

[tool.hatch.envs.testing.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]

[[tool.hatch.envs.all.matrix]]
python = ["3.10", "3.11", "3.12"]

[tool.hatch.envs.types]
dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/backend tests}"

[tool.coverage.run]
source_pkgs = ["backend", "tests"]
branch = true
parallel = true
omit = [
  "src/backend/__about__.py",
]

[tool.coverage.paths]
backend = ["src/backend", "*/backend/src/backend"]
tests = ["tests", "*/backend/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
